rename trait -> traits
